requirements.txt
streamlit
Pillow
langchain_google_genai
python-dotenv
tensorflow
torch